update: enforce ts extensions and add webm demuxer
