+++ src/playback/processing/streamProcessor.ts
@@ -48,7 +48,7 @@ import type {
 } from '../../typings/streamProcessor.types.ts'
 import { logger } from '../../utils.js'
 import FlvDemuxer from '../demuxers/Flv.js'
-import WebmOpusDemuxer from '../demuxers/WebmOpus.js'
+import WebmOpusDemuxer from '../demuxers/WebmOpus.ts'
 import { Decoder as OpusDecoder, Encoder as OpusEncoder } from '../opus/Opus.js'
 import { RingBuffer } from '../structs/RingBuffer.js'
 import { FadeTransformer } from './FadeTransformer.js'
