+++ src/workers/main.ts
@@ -3,7 +3,7 @@ import os from 'node:os'
 import { monitorEventLoopDelay } from 'node:perf_hooks'
 import v8 from 'node:v8'
 
-import { GatewayEvents } from '../constants'
+import { GatewayEvents } from '../constants.ts'
 import ConnectionManager from '../managers/connectionManager.js'
 import CredentialManager from '../managers/credentialManager.js'
 import LyricsManager from '../managers/lyricsManager.js'
@@ -13,7 +13,7 @@ import RoutePlannerManager from '../managers/routePlannerManager.js'
 import SourceManager from '../managers/sourceManager.js'
 import StatsManager from '../managers/statsManager.js'
 import TrackCacheManager from '../managers/trackCacheManager.js'
-import { Player } from '../playback/player'
+import { Player } from '../playback/player.ts'
 import { createPCMStream } from '../playback/processing/streamProcessor.ts'
 import { bufferPool } from '../playback/structs/BufferPool.js'
 import type { TrackInfoExtended } from '../typings/player.types.ts'
